### 2018-01-28 #554 @whedon
[PRE REVIEW]: TSrepr R package: Time Series Representations
**Submitting author:** @PetoLau (<a href="http://orcid.org/0000-0002-3501-8783">Peter Laurinec</a>)
**Repository:** <a href="https://github.com/PetoLau/TSrepr" target ="_blank">https://github.com/PetoLau/TSrepr</a>
**Version:** v1.0.0
