+++ vignettes/papers.bib
@@ -11,7 +11,7 @@
   author={Laurinec, Peter and Luck{\'a}, M{\'a}ria},
   booktitle={Lecture Notes in Engineering and Computer Science: Proceedings of The World Congress on Engineering and Computer Science 2016},
   pages={458--463},
-  month = {oct},  
+  month = {oct},
   year={2016}
 }
 @article{Esl,
@@ -56,4 +56,12 @@ year = {2006}
   pages={297--301},
   year={1965},
   publisher={JSTOR}
+}
+@inproceedings{ratanamahatana,
+  title={A novel bit level time series representation with implication of similarity search and clustering},
+  author={Ratanamahatana, Chotirat and Keogh, Eamonn and Bagnall, Anthony J and Lonardi, Stefano},
+  booktitle={Pacific-Asia Conference on Knowledge Discovery and Data Mining},
+  pages={771--777},
+  year={2005},
+  organization={Springer}
 }
