### 2018-02-04 #554 @arokem
Sure. I can edit this.

Is this one of these cases that should also be reviewed by rOpenSci in parallel?
