### 2018-02-04 #554 @arokem
👋 @kellieotto : would you like to review this submission for JOSS?
