### 2018-02-19 PetoLau/TSrepr#2 @kellieotto
Missing unit tests
This is detailed feedback for your JOSS review openjournals/joss-reviews#577.

R packages should have a `tests/` directory containing unit tests. [Hadley Wickham’s R packages book](http://r-pkgs.had.co.nz/tests.html) has a whole chapter explaining how to do this.

Better yet, you could set up a continuous integration system with [Travis CI](https://travis-ci.org) and add a badge to your README.md that shows that your package passes all of the tests.
