### 2018-02-19 PetoLau/TSrepr#3 @kellieotto
Community guidelines
This is detailed feedback for your JOSS review openjournals/joss-reviews#577.

There’s an email contact for questions and comments, but no guidelines for contribution. I would suggest adding a CONTRIBUTING.md file to the repository. I browsed through some ROpenSci packages and many of them have guidelines like this. For example:

- [Tabulizer](https://github.com/ropensci/tabulizer/blob/master/CONTRIBUTING.md)
- [Rentrez](https://github.com/ropensci/rentrez/blob/f815d024242d873cee8e2105be17b2ca02ea3ec1/.github/CONTRIBUTING.md)
- [Rorcid](https://github.com/ropensci/rorcid/blob/d4ec96a7af0f2787351312af26122da657837b8a/.github/CONTRIBUTING.md)
