### 2018-02-19 PetoLau/TSrepr#4 @kellieotto
Paper needs to be proofread
This is detailed feedback for your JOSS review openjournals/joss-reviews#577.

The paper accompanying the software is unclear because it needs to be proofread for grammar and correct wording. I can give more detailed feedback if needed.
