### 2018-02-19 #577 @kellieotto
More detailed comments are available for the author in issues in the package's repository.

- Unit tests are absent (PetoLau/TSrepr#2)
- Community guidelines for contribution should be added (PetoLau/TSrepr#3)
- Paper needs proofreading (PetoLau/TSrepr#4)
