+++ CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# CONTRIBUTING #
+
+### Please contribute!
+
+I love collaboration.
+
+### Bugs?
+
+* Submit an issue on the [Issues page](https://github.com/PetoLau/TSrepr/issues).
+
+### Code contributions
+
+* Fork this repo to your Github account,
+* Clone your version on your account down to your machine from your account, e.g,. `git clone https://github.com/<yourgithubusername>/TSrepr.git`
+* Make sure to track progress upstream (i.e., on our version of `TSrepr` at `petolau/TSrepr`) by doing `git remote add upstream https://github.com/petolau/TSrepr.git`. Before making changes make sure to pull changes in from upstream by doing either `git fetch upstream` then merge later or `git pull upstream` to fetch and merge in one step,
+* Make your changes (bonus points for making changes on a new feature branch),
+* Push up to your account,
+* Submit a pull request to home base at `petolau/TSrepr`.
+
+### Also, check my blog for more info [petolau.github.io](https://petolau.github.io/)
+
+### Prefer to Email? Get in touch: [tsreprpackage@gmail.com](mailto:tsreprpackage@gmail.com)
+
+### Thanks for contributing!
