### 2018-02-23 PetoLau/TSrepr#3 @PetoLau
I added CONTRIBUTING.md file to main directory (inspired by Rorcid package).
Thank you!
c22993d
