### 2018-02-23 402e9ac
#2 Add: test folder and travis.yml file
