+++ .travis.yml
@@ -1,0 +1,18 @@
+# Sample .travis.yml for R projects
+
+language: r
+sudo: false
+cache: packages
+
+r:
+  - oldrel
+  - release
+  - devel
+
+env:
+ global:
+   - CRAN: http://cran.rstudio.com
+
+notifications:
+  email:
+    on_failure: change
