### 2018-02-23 31e7e44
#2 Add: test folder and travis.yml file
