### 2018-02-23 16b75b4
Update: readme travis and contributions
