+++ README.Rmd
@@ -14,6 +14,7 @@ knitr::opts_chunk$set(
 
 # TSrepr
 
+[![Travis-CI Build Status](https://travis-ci.org/PetoLau/TSrepr.svg?branch=master)](https://travis-ci.org/PetoLau/TSrepr)
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/TSrepr)](https://cran.r-project.org/package=TSrepr)
 [![Downloads](http://cranlogs.r-pkg.org/badges/TSrepr)](https://cran.r-project.org/package=TSrepr)
 [![Downloads](http://cranlogs.r-pkg.org/badges/grand-total/TSrepr?color=green)](https://cran.r-project.org/package=TSrepr)
@@ -96,4 +97,5 @@ ggplot(data_plot, aes(time, value, color = type, size = type)) +
 
 ## Contact
 
-For any suggestions and comments write me an email at: [tsreprpackage@gmail.com](mailto:tsreprpackage@gmail.com)
+ - For any suggestions and comments write me an email at: [tsreprpackage@gmail.com](mailto:tsreprpackage@gmail.com).
+ - For contribution options, check CONTRIBUTING.md file, please.
