### 2018-02-23 094eb52
Add: first simple test
