### 2018-02-23 52b9a4b
Add: first simple test
