### 2018-02-23 fa9e5f8
#2 Add: test R files for repr_..() functions
