### 2018-02-23 14ee4fe
#2 Add: test R files for repr_..() functions
