### 2018-02-23 30fd816
Update: if else simplifying in pip and pla
