### 2018-02-23 058a037
Update: if else simplifying in pip and pla
