### 2018-02-23 PetoLau/TSrepr#2 @PetoLau
@kellieotto I added `testthat` testings for all `repr_...()` functions.
I also created Travis CI checking for this project.
fa9e5f8
