+++ .travis.yml
@@ -8,6 +8,12 @@ r:
   - oldrel
   - release
 
+r_github_packages:
+  - r-lib/covr
+
+after_success:
+  - Rscript -e 'covr::codecov()'
+
 notifications:
   email:
     on_failure: change
