### 2018-03-04 4c25cb1
#2 Update: more tests based on covr
