### 2018-03-04 af06c54
#2 Update: more tests based on covr
