### 2018-03-05 86ea775
Update: tests for src/ functions
