### 2018-03-05 8c2e878
Update: tests for src/ functions
