+++ DESCRIPTION
@@ -4,7 +4,8 @@ Title: Time Series Representations
 Version: 1.0.0.900
 Date: 2018-01-26
 Authors@R: person("Peter", "Laurinec", email = "tsreprpackage@gmail.com",
-  role = c("aut", "cre"))
+  role = c("aut", "cre"), 
+  comment = c(ORCID = "0000-0002-3501-8783"))
 Description: Methods for representations (i.e. dimensionality reduction, preprocessing, feature extraction) of time series to help more accurate and effective time series data mining.
     Non-data adaptive, data adaptive, model-based and data dictated (clipped) representation methods are implemented. Also min-max and z-score normalisations, and forecasting accuracy measures are implemented.
 License: GPL-3 | file LICENSE
