### 2018-03-07 984e902
#4 Update: paper proofread and updated
