+++ paper/paper.bib
@@ -20,12 +20,20 @@ Dicretization, i.e. SAX},
     note = {R package version 1.0.2.900},
     url = {https://CRAN.R-project.org/package=jmotif},
 }
-@inproceedings{Lau,
+@inproceedings{Lau1,
   title={Comparison of Representations of Time Series for Clustering Smart Meter Data},
   author={Laurinec, Peter and Luck{\'a}, M{\'a}ria},
   booktitle={Lecture Notes in Engineering and Computer Science: Proceedings of The World Congress on Engineering and Computer Science 2016},
   pages={458--463},
   year={2016}
+}
+@inproceedings{Lau2,
+  title={Adaptive Time Series Forecasting of Energy Consumption Using Optimized Cluster Analysis},
+  author={Laurinec, Peter and L{\'o}derer, Marek and Vrablecov{\'a}, Petra and Luck{\'a}, M{\'a}ria and Rozinajov{\'a}, Viera and Ezzeddine, Anna Bou},
+  booktitle={Data Mining Workshops (ICDMW), 2016 IEEE 16th International Conference on},
+  pages={398--405},
+  year={2016},
+  organization={IEEE}
 }
 @article{rcpp,
    author = {Dirk Eddelbuettel and Romain Francois},
@@ -38,4 +46,15 @@ Dicretization, i.e. SAX},
    pages = {1--18},
    doi = {10.18637/jss.v040.i08},
    url = {https://www.jstatsoft.org/v040/i08}
+}
+@article{Esling,
+author = {Esling, Philippe and Agon, Carlos},
+journal = {ACM Computing Surveys},
+keywords = {Distance measures,data indexing,data mining,query by content,sequence matching,similarity measures,stream analysis,temporal analysis,time series},
+number = {1},
+pages = {1--34},
+publisher = {ACM},
+title = {{Time-series data mining}},
+volume = {45},
+year = {2012}
 }
