### 2018-03-07 057431f
#4 Update: paper proofread and updated
