+++ paper/paper.md
@@ -11,27 +11,29 @@ affiliations:
 - name: Faculty of Informatics and Information Technologies,
     Slovak University of Technology in Bratislava
   index: 1
-date: 27 January 2018
+date: 7 March 2018
 bibliography: paper.bib
 output: pdf_document
 ---
 
 # Summary
 
-TSrepr [@tsrepr] is R package for time series representations computing. Time series representations are, in other words, methods for dimensionality reduction, feature extraction or preprocessing of time series. They are used for:
+`TSrepr` [@tsrepr] is an R package for time series representations computing. Time series representations are, in other words, methods for dimensionality reduction, feature extraction or for the preprocessing of time series. They are used for [@Esling]:
 
- * significant reduction of the time series dimensionality
- * emphasis on fundamental (essential) shape characteristics
- * implicit noise handling
- * reducing the dimension will reduce memory requirements and computational complexity of consequent machine learning methods (classification or clustering).
+ * Significant reduction of the time series dimensionality,
+ * Emphasis of fundamental (essential) shape characteristics,
+ * Implicit noise handling,
+ * Dimension reduction will reduce the memory requirements and computational complexity of consequent machine learning methods (e.g., classification or clustering).
+ 
+The `TSrepr` package contains various methods and types of time series representations including the Piecewise Aggregate Approximation (PAA), the Discrete Fourier Transform (DFT), the Perceptually Important Points (PIP), the Symbolic Aggregate approXimation (SAX), the Piecewise Linear Approximation (PLA) and Clipping. Except for these well-known methods, additional methods suitable for seasonal time series are implemented. These methods are based on the model, for example multiple linear regression, robust regression, generalised additive model or triple exponential smoothing [@Lau1, @Lau2]. Own developed feature extraction methods from the Clipping representation are also implemented - FeaClip and FeaTrend.
+In Figure 1, the comparison of all eight available model-based representations in the `TSrepr` on electricity consumption time series from the randomly picked residential consumer is shown.
 
-The TSrepr package contains various methods and types of time series representations including Piecewise Aggregate Approximation (PAA), Discrete Fourier Transform (DFT), Perceptually Important Points (PIP), Symbolic Aggregate approXimation (SAX), Piecewise Linear Approximation (PLA) or Clipping. Except for these well-known methods, additional methods suitable for seasonal time series are implemented. These methods are based on the model as for example multiple linear regression, robust regression, generalized additive model or triple exponential smoothing [@Lau]. Own feature extraction methods from Clipping representation are also implemented - FeaClip and FeaTrend.
+![](modelBased.pdf)
+Figure 1. The comparison of model-based time series representations on electricity consumption time series. The length of representations is $48$, the same as frequency of the daily season of the used time series.
 
-Additional useful functions and methods related to time series representations were also implemented. Package includes functions for normalisations and denormalisations of time series - z-score and min-max methods. It supports simple computation of windowing method, a matrix of representations and forecasting accuracy measures (MAE, RMSE, MAPE, sMAPE etc.).
-
+Additional useful functions and methods related to time series representations were also implemented. The `TSrepr` package includes functions for normalisations and denormalisations of time series - z-score and min-max methods. It supports the simple computation of the windowing method, a matrix of representations and forecasting accuracy measures (MAE, RMSE, MAPE, sMAPE etc.).
 Methods (functions) were implemented in base R and also in C++ for fast computations. In R, C++ programmes can be written thanks to the package Rcpp [@rcpp].
 
-The general package for time series representations computations missed until now.
[…]
