### 2018-03-07 #577 @PetoLau
Hello!
Thank you very much for your feedback.
I tried to solve all problems:

 - unit tests for all functions in my project was created (codecov 93%) and Travis CI was deployed,
 - CONTRIBUTING.md file was added for the clarification of possible contributions,
 - paper was proofread.

I am looking forward for your response.
