### 2018-03-08 PetoLau/TSrepr#4 @kellieotto
Much better. I would suggest one change. In the last paragraph, I think the grammar is not quite right in the first sentence. I would say "So far, no general package for time series representations computations has been created."
