+++ README.Rmd
@@ -99,4 +99,4 @@ ggplot(data_plot, aes(time, value, color = type, size = type)) +
 ## Contact
 
  - For any suggestions and comments write me an email at: [tsreprpackage@gmail.com](mailto:tsreprpackage@gmail.com).
- - For contribution options, check CONTRIBUTING.md file, please.
+ - For contribution options, check [CONTRIBUTING.md](https://github.com/PetoLau/TSrepr/blob/master/CONTRIBUTING.md) file, please.
