### 2018-03-11 #577 @arokem
@PetoLau : Everything looks good!

The next step would be for you to create a new release of the software and archive it (e.g., on [Zenodo](https://zenodo.org)). Once you have done that, please post the DOI of the software here, and we will go ahead with the publication of the article (which will include this DOI).
